JC- Fixed ship movement and shooting. Tried fixing scoreboard

+++ GD4SFMLGameWorld/Player.hpp
@@ -42,7 +42,7 @@ private:
 	std::map<ActionID, Command> mActionBinding;
 	std::map<ActionID, bool> mActionProxies;
 	MissionStatusID mCurrentMissionStatus;
-	sf::Uint8 mIdentifier;
+	sf::Int32 mIdentifier;
 	sf::TcpSocket* mSocket;
 	PlayerID mType;
 	sf::Uint8 mScore;
